Move HeaderDrawer and HeaderHotkeys inside of the Header section
