+++ src/sections/Header/HeaderHotkeys/useHotkeysConfig.ts
